Solver atualizado para rodar localmente "rápido" Heurística atualizada e agora funciona Testes realizados e resultados salvos

+++ ADMAG.hpp
@@ -11,19 +11,19 @@
 #include "candidato.hpp"
 
 // Variáveis globais definidas na inicialização do programa baseado no tamanho de uma solução
-extern uint16_t tam_particao;
-extern uint8_t max_mutacoes;
-extern uint32_t total_pais;
+extern unsigned int tam_particao;
+extern unsigned int max_mutacoes;
+extern unsigned int total_pais;
 
-candidato criarCandidatoAleatorio(const std::vector<uint8_t>& p, std::mt19937& gen);
+candidato criarCandidatoAleatorio(const std::vector<unsigned int>& p, std::mt19937& gen);
 
-vetorDeCandidatos selecionarPais(const vetorDeCandidatos& populacao_atual, const uint16_t k, const uint32_t limite_populacional, std::mt19937& gen);
+vetorDeCandidatos selecionarPais(const vetorDeCandidatos& populacao_atual, const unsigned int k, const unsigned int limite_populacional, std::mt19937& gen);
 
-vetorDeCandidatos gerarFilhos(const vetorDeCandidatos& pais, const uint16_t limite_populacional, std::mt19937& gen);
+vetorDeCandidatos gerarFilhos(const vetorDeCandidatos& pais, const unsigned int limite_populacional, std::mt19937& gen);
 
-void causarMutacoes(std::vector<uint8_t>& genes, std::mt19937& gen);
+void causarMutacoes(std::vector<unsigned int>& genes, std::mt19937& gen);
 
-vetorDeCandidatos determinarNovaPopulacao(const vetorDeCandidatos& populacao_atual, const vetorDeCandidatos& novos_candidatos, const uint16_t max_atuais, const uint16_t max_novos);
+vetorDeCandidatos determinarNovaPopulacao(const vetorDeCandidatos& populacao_atual, const vetorDeCandidatos& novos_candidatos, const unsigned int max_atuais, const unsigned int max_novos);
 
 // Função principal do algoritmo
-candidato ADMAG(const uint16_t k, const uint32_t limite_populacional, const uint16_t max_atuais, const uint16_t max_novos, const bool modo_temporal, const uint32_t condicao_de_parada, const std::vector<uint8_t>& solucao_inicial, const int32_t seed);
+candidato ADMAG(const unsigned int k, const unsigned int limite_populacional, const unsigned int max_atuais, const unsigned int max_novos, const bool modo_temporal, const unsigned int condicao_de_parada, const std::vector<unsigned int>& solucao_inicial, const int seed);
